requirements.txt
flask==3.1.3
gunicorn>=23.0.0; sys_platform != "win32"
waitress>=3.0.0; sys_platform == "win32"
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
python-dotenv>=1.0.0
jinja2>=3.1.0
